restructured src folder and cleaned code
