--- src/app/posts/posts.component.scss
@@ -1,3 +1,0 @@
-h2 {
-  color: blue;
-}
