+++ src/main.ts
@@ -1,7 +1,7 @@
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import { AppModule } from './app/app.module';
+import { AppModule } from './modules/app/app.module';
 import { environment } from './environments/environment';
 
 if (environment.production) {
