+++ src/modules/app/app-routing.module.ts
@@ -1,13 +1,13 @@
 import { NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
 import { HomeComponent } from "./home/home.component";
-import { PostsComponent } from "./posts/posts.component";
+import { TasksComponent } from "./tasks/tasks.component";
 import { SignInComponent } from "./sign-in/sign-in.component";
 import { SignUpComponent } from "./sign-up/sign-up.component";
 
 const routes: Routes = [
   {path: '', component: HomeComponent},
-  {path: 'posts', component: PostsComponent},
+  {path: 'tasks', component: TasksComponent},
   {path: 'sign-in', component: SignInComponent},
   {path: 'sign-up', component: SignUpComponent}
 ]
