+++ src/modules/app/app.component.html
@@ -1,7 +1,7 @@
 <nav class='navbar'>
   <ul>
     <li><a routerLink='/' routerLinkActive='active' [routerLinkActiveOptions]='{exact: true}'>Home</a></li>
-    <li><a routerLink='/posts' routerLinkActive='active'>Posts</a></li>
+    <li><a routerLink='/tasks' routerLinkActive='active'>Tasks</a></li>
     <li><a routerLink='/sign-in' routerLinkActive='active'>SignIn</a></li>
     <li><a routerLink='/sign-up' routerLinkActive='active'>SignUp</a></li>
   </ul>
