+++ src/modules/app/app.component.ts
