+++ src/modules/app/sign-in/sign-in.component.html
@@ -5,13 +5,13 @@
       <form [formGroup]='form' (ngSubmit)='onSubmit()'>
         
           <mat-form-field class="example-full-width" appearance="fill">
-            <mat-label>Email</mat-label>
+            <mat-label>Email*</mat-label>
             <input matInput type="text" placeholder="pat@example.com" formControlName="email" (blur)="getEmailErrorMessage()">
             <mat-error>{{emailError}}</mat-error>
           </mat-form-field>
         
           <mat-form-field class="example-full-width" appearance="fill">
-            <mat-label>Password</mat-label>
+            <mat-label>Password*</mat-label>
             <input matInput type="password" formControlName="password" (blur)="getPasswordErrorMessage()">
             <mat-error>{{passwordError}}</mat-error>
           </mat-form-field>
