+++ src/modules/app/sign-up/sign-up.component.scss
