+++ src/modules/app/sign-up/sign-up.component.ts
@@ -1,6 +1,6 @@
-import { Component, OnInit } from "@angular/core";
-import {FormControl, FormGroup, Validators} from '@angular/forms';
-import { Router } from "@angular/router";
+import { Component, OnInit } from '@angular/core';
+import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-sign-up',
@@ -27,7 +27,10 @@ export class SignUpComponent implements OnInit {
         Validators.required,
         Validators.minLength(3),
       ]),
-      email: new FormControl('', [Validators.email, Validators.required]),
+      email: new FormControl('', [
+        Validators.email, 
+        Validators.required
+      ]),
       password: new FormControl('', [
         Validators.required,
         Validators.minLength(8),
@@ -42,37 +45,31 @@ export class SignUpComponent implements OnInit {
   getFirstNameErrorMessage(): void {
     this.firstnameError = this.form.value.firstname.trim()
       ? 'Not a valid First Name'
-      : 'You must enter a value';
+      : 'You must enter a value'
   }
 
   getLastNameErrorMessage(): void {
     this.lastnameError = this.form.value.lastname.trim()
       ? 'Not a valid Last Name'
-      : 'You must enter a value';
+      : 'You must enter a value'
   }
 
   getEmailErrorMessage(): void {
-    if (this.form.value.email.trim()) {
-      this.emailError = 'Not a valid Email';
-    } else {
-      this.emailError = 'You must enter a value';
-    }
+    this.emailError = this.form.value.email.trim()
+      ? 'Not a valid Email'
+      : 'You must enter a value'
   }
 
   getPasswordErrorMessage(): void {
-    if (this.form.value.password.trim()) {
-      this.passwordError = 'Not a valid password';
-    } else {
-      this.passwordError = 'You must enter a value';
-    }
+    this.passwordError = this.form.value.password.trim()
+      ? 'Not a valid Password'
+      : 'You must enter a value'
   }
 
   getConfirmPasswordErrorMessage(): void {
-    if (this.form.value.confirmpassword !== this.form.value.password) {
-      this.confirmPasswordError = 'Value must be matched with password';
-    } else {
-      this.confirmPasswordError = 'You must enter a value';
-    }
+    this.confirmPasswordError = this.form.value.confirmpassword !== this.form.value.password
+      ? 'Value must be matched with Password'
+      : 'You must enter a correct value of Password'
   }
 
   goToSignInPage(): void {
@@ -80,7 +77,9 @@ export class SignUpComponent implements OnInit {
   }
 
   onSubmit(): void {
-    if (this.form.valid) {
+    if (this.form.value.confirmpassword !== this.form.value.password) {
+      this.getConfirmPasswordErrorMessage();
+    } else if(this.form.valid && this.form.value.confirmpassword === this.form.value.password) {
       console.log({ ...this.form.value });
 
       this.form.reset();
