+++ src/modules/app/tasks/tasks.component.html
@@ -1,3 +1,3 @@
 <div class='page_body'>
-  <h1>Posts Page!</h1>
+  <h1>Tasks Page!</h1>
 </div>
